+++ quarkus/tests/integration/src/main/java/org/junit/rules/TestRule.java
@@ -1,0 +1,6 @@
+package org.junit.rules;
+
+// WORKAROUND: https://github.com/testcontainers/testcontainers-java/issues/970#issuecomment-625044008
+@SuppressWarnings("unused")
+public interface TestRule {
+}
